remove dummy net; delete two files
