+++ test/discrete/test_a2c_with_il.py
@@ -10,11 +10,7 @@ from tianshou.env import VectorEnv
 from tianshou.data import Collector, ReplayBuffer
 from tianshou.policy import A2CPolicy, ImitationPolicy
 from tianshou.trainer import onpolicy_trainer, offpolicy_trainer
-
-if __name__ == '__main__':
-    from net import Net, Actor, Critic
-else:  # pytest
-    from test.discrete.net import Net, Actor, Critic
+from tianshou.utils.net.discrete import Net, ActorHead, CriticHead
 
 
 def get_args():
@@ -67,8 +63,8 @@ def test_a2c_with_il(args=get_args()):
     test_envs.seed(args.seed)
     # model
     net = Net(args.layer_num, args.state_shape, device=args.device)
-    actor = Actor(net, args.action_shape).to(args.device)
-    critic = Critic(net).to(args.device)
+    actor = ActorHead(net, args.action_shape).to(args.device)
+    critic = CriticHead(net).to(args.device)
     optim = torch.optim.Adam(list(
         actor.parameters()) + list(critic.parameters()), lr=args.lr)
     dist = torch.distributions.Categorical
@@ -111,7 +107,7 @@ def test_a2c_with_il(args=get_args()):
     if args.task == 'CartPole-v0':
         env.spec.reward_threshold = 190  # lower the goal
     net = Net(1, args.state_shape, device=args.device)
-    net = Actor(net, args.action_shape).to(args.device)
+    net = ActorHead(net, args.action_shape).to(args.device)
     optim = torch.optim.Adam(net.parameters(), lr=args.il_lr)
     il_policy = ImitationPolicy(net, optim, mode='discrete')
     il_test_collector = Collector(il_policy, test_envs)
